requirements.txt
Django==3.2.8
psycopg2-binary==2.9.1
requests==2.26.0
ibm-watson==5.3.0
django-environ==0.7.0
